requirements.txt
langchain>=0.2.12
langchain-openai>=0.2.0
openai>=1.52.0
tiktoken>=0.7.0
pytest>=8.2.0
pytest-cov>=5.0.0
python-dotenv>=1.0.1
rich>=13.7.1
